+++ lib/faktum/include/faktum/storage.hpp
@@ -104,7 +104,7 @@ public:
 
 	virtual ~FStorage();
 
-	static void* __cdecl operator new(unsigned int, FObject*, const FName&, const FName&, unsigned int);
+	static void* __cdecl operator new(size_t, FObject*, const FName&, const FName&, unsigned int);
 
 	FStorage& operator=(const FStorage& other);
