+++ lib/faktum/include/faktum/text.hpp
@@ -23,7 +23,7 @@ public:
 	const FString& operator=(const wchar_t* str);
 	const FString& operator=(wchar_t c);
 
-	static void* __cdecl operator new(unsigned int, FObject*, const FName&, const FName&, unsigned int);
+	static void* __cdecl operator new(size_t, FObject*, const FName&, const FName&, unsigned int);
 
 	virtual void Serialize(FArchive& arc) override;
